+++ src/processor/connector/python/adapter.py
@@ -4,7 +4,7 @@ from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
 
 def execute_query(query: str, cred: dict[str, str]):
   try:
-    conn = open_connection(cred)
+    conn = psycopg2.connect(cred['conn_string'])
     conn.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
     cursor = conn.cursor()
     cursor.execute(query)
@@ -17,19 +17,3 @@ def execute_query(query: str, cred: dict[str, str]):
     if 'conn' in locals():
       conn.close()
       print('Connection closed.')
-
-def open_connection(cred: dict[str, str]):
-  # connector = {
-  #   'host': cred['host'],
-  #   'port': cred['port'],
-  #   'database': cred['database'],
-  #   'sslmode': cred['sslmode'],
-  #   'user': cred['user'],
-  #   'password': cred['password'],
-  # }
-
-  print(f"conn_string = {cred['conn_string']}")
-
-  conn = psycopg2.connect(cred['conn_string'])
-  # print(f"Info: Connected to '{connector['host']}:{connector['port']}/{connector['database']}' with user '{connector['user']}'")
-  return conn